コマンド登録のインターフェイスを作成
