+++ src/models/bots/commands/index.ts
@@ -1,0 +1,4 @@
+// コマンドを作製したらここに並べていくことで登録完了
+import './paizaIO';
+import './ping';
+import './moja';
