+++ src/models/bots/commands/ping.ts
@@ -1,0 +1,6 @@
+import { CommandBot } from '../commandBot';
+
+// test command
+CommandBot.register('/ping', async () => {
+  return 'pong!';
+});
