+++ src/models/bots/index.ts
@@ -1,2 +1,5 @@
 export * from './timeLineBot';
 export * from './commandBot';
+
+// ここで import することで commands 以下で CommandBot を利用してもよいことになる
+import './commands';
